calculator-project/requirements.txt
pywebview
flask
flask_cors